app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.2
certifi==2023.11.17
click==8.1.7
dill==0.3.8
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
orjson==3.9.12
passlib==1.7.4
platformdirs==4.1.0
psutil==5.9.8
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pylint==3.0.3
python-dotenv==1.0.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
